requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
python-socketio>=5.8.0
Flask>=2.3.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0
werkzeug>=2.3.0
